+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased](https://github.com/jshwi/jss/compare/v0.1.0...HEAD)
 ------------------------------------------------------------------------
 ### Added
+- Adds follow functionality
 - Adds edit profile page
 - Records user's last login time
 - Adds page to display individual posts